Added CODE_OF_CONDUCT.md
docs: add a Code of Conduct file #31

Added a code of conduct based on a template available at GitHub. The Covenant Contributor model was chosen based on the size of the project and community.

Added Micha's GitHub link as the contact for code enforcement. If the contact person doesn't agree with sharing their contact, we can change to another person.